Move hash_token and _hash_token_if_needed to _types.py
so that I can use `_hash_token_if_needed` in
`UserAPIKeyAuth.check_api_key` in `_types.py`.

addresses
[comment](https://github.com/BerriAI/litellm/pull/10601#discussion_r2076728733)
from @ishaan-jaff
